docs(repeat-till): Reword top-level doc string
The old docstring usees the arbitrary variables `f` and `g` to
differentiate between the two types of parsers, even though the
variables have never been established.

This change replaces those variables with the actual parameter names
that're passed to `repeat_till`, which should make this much more
intuitive to read.